Fix sfAutoload and sfCommandApplication

+++ lib/autoload/sfAutoload.class.php
@@ -129,7 +129,10 @@ class sfAutoload
       self::$freshCache = false;
       if ($force)
       {
-        unlink($configuration->getConfigCache()->getCacheName('config/autoload.yml'));
+        if (file_exists($configuration->getConfigCache()->getCacheName('config/autoload.yml')))
+        {
+          unlink($configuration->getConfigCache()->getCacheName('config/autoload.yml'));
+        }
       }
     }
